pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "inception-external-recommender"
version = "0.1.0-dev"
description = "INCEpTION external recommender library in Python"
readme = "README.md"
requires-python = ">=3.9.0"
license = {text = "Apache License 2.0"}
authors = [
    {name = "The INCEpTION team", email = "inception-users@googlegroups.com"}
]
keywords = ["uima", "dkpro", "inception", "nlp"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
    "Topic :: Scientific/Engineering :: Human Machine Interfaces",
    "Topic :: Text Processing :: Linguistic",
]

dependencies = [
    "flask",
    "filelock",
    "dkpro-cassis>=0.9.1",
    "joblib",
    "gunicorn",
    "deprecation",
    "scikit-learn>=1.0.1",
    "sklearn_crfsuite~=0.3.6",
    "numpy>=1.24.0,<3.0.0",
    "huggingface_hub>=0.23.0",
    "spacy>=3.7.0",
    "nltk~=3.5",
    "jieba~=0.42",
    "sentence-transformers>=3.0.0",
    "lightgbm~=4.2.0",
    "diskcache~=5.2.1",
    "simalign~=0.4",
    "flair>=0.13.1",
    "transformers[torch]>=4.41.0",
    "pip",  # Required for spacy model auto-download during tests
]

[dependency-groups]
dev = [
    "ruff",
    "wget",
    "tox",
    "pytest",
    "codecov",
    "pytest-cov",
    "taskipy",
]
doc = [
    "sphinx",
    "sphinx-autodoc-typehints",
    "sphinx-rtd-theme",
]

[project.urls]
Homepage = "https://inception-project.github.io/"
"Bug Tracker" = "https://github.com/inception-project/inception-external-recommender/issues"
Documentation = "https://github.com/inception-project/inception-external-recommender"
"Source Code" = "https://github.com/inception-project/inception-external-recommender"

[tool.hatch.build.targets.wheel]
packages = ["ariadne"]

[tool.taskipy.tasks]
test = "pytest -m 'not performance' tests/"
test-cov = "pytest -m 'not performance' --cov=./ --cov-report=xml tests/"
serve = "gunicorn -w 4 -b 127.0.0.1:5000 --reload wsgi:app"
format = "ruff format ariadne/ scripts/ tests/ wsgi.py"
lint = "ruff check ariadne/ scripts/ tests/ wsgi.py"
fix = "ruff check --fix ariadne/ scripts/ tests/ wsgi.py"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
# Equivalent to flake8's E9,F63,F7,F82 (syntax errors and undefined names)
select = [
    "E9",   # Runtime errors
    "F63",  # Invalid print syntax
    "F7",   # Syntax errors in type comments
    "F82",  # Undefined names in __all__
    "F",    # Pyflakes (includes undefined names, unused imports, etc.)
]
# Equivalent to flake8's --max-complexity=10
mccabe = { max-complexity = 10 }

[tool.ruff.lint.per-file-ignores]
# Allow star imports in test files for test utilities
"tests/**/*.py" = ["F403", "F405"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
